backend/requirements.txt
flask==3.0.2
flask-cors==4.0.0
flask-sqlalchemy==3.1.1
python-dotenv==1.0.1
openai-whisper==20231117
numpy==1.26.4
librosa==0.10.2.post1
soundfile==0.12.1
